Refactor image
