+++ src/image/image.pyi
