+++ src/image/image_doc.h
